Django-Middleware-0x03/pyproject.toml
[project]
name = "messaging-app-middleware"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-mysql>=4.19.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "gunicorn>=23.0.0",
    "mysqlclient>=2.2.7",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "rabbitmq>=0.2.0",
    "ruff>=0.14.5",
    "sqlparse>=0.5.3",
]
